farmacia_python/requirements.txt
flask==3.1.0
supabase==2.27.2
python-dotenv==1.2.1
flask-cors==3.0.0
reportlab==4.4.0
groq==1.0.0
PyJWT==2.8.0
gunicorn==22.0.0
langchain==0.1.20
langchain-openai==0.1.14
openai==1.42.0
pytest==8.2.2
pytest-flask==1.3.0